Add new route constants and middleware for improved API and web structure
- Introduced new route constants for various resources, enhancing the organization and maintainability of route definitions.
- Added middleware for campus selection checks, ensuring users are directed to the appropriate campus context.
- Created new seeder classes for initial setup, streamlining the database seeding process for academic structures and user accounts.
- Updated existing routes to reflect the new structure, improving clarity and consistency across the application.
- Removed deprecated routes and tests to clean up the codebase and improve overall performance.
- Enhanced documentation to reflect the new routing and middleware capabilities.

+++ app/Models/Lecture.php
@@ -193,7 +193,7 @@ class Lecture extends Model
 
     public function getYearsOfServiceAttribute(): int
     {
-        return $this->hire_date ? $this->hire_date->diffInYears(now()) : 0;
+        return (int) ($this->hire_date ? $this->hire_date->diffInYears(now()) : 0);
     }
 
     public function getIsContractActiveAttribute(): bool
